+++ src/main/java/convention/action/users/DeleteUserAction.java
@@ -1,9 +1,9 @@
-package convention.action;
+package convention.action.users;
 
 import ar.com.kfgodel.appbyconvention.operation.api.ApplicationOperation;
 import ar.com.kfgodel.dependencies.api.DependencyInjector;
 import ar.com.kfgodel.orm.api.operations.basic.DeleteById;
-import convention.persistent.Procedure;
+import convention.persistent.Usuario;
 import convention.rest.api.tos.IdReferenceTo;
 
 import javax.annotation.Resource;
@@ -11,16 +11,17 @@ import javax.inject.Inject;
 import java.util.function.Function;
 
 /**
- * This type represents the frontend action to delete a procedure
- * Created by kfgodel on 14/11/16.
+ * This type represents the frontend action for removing a user
+ * <p>
+ * Created by kfgodel on 17/11/16.
  */
-@Resource(name = "DELETE/procedure")
-public class DeleteProcedure implements Function<IdReferenceTo, Void> {
+@Resource(name = "DELETE/user")
+public class DeleteUserAction implements Function<IdReferenceTo, Void> {
   @Override
   public Void apply(IdReferenceTo reference) {
     createOperation()
       .insideATransaction()
-      .apply(DeleteById.create(Procedure.class, reference.getId()));
+      .apply(DeleteById.create(Usuario.class, reference.getId()));
     return null;
   }
